gourmet/requirements.txt
bcrypt
email-validator
fastapi
psycopg2-binary
python-jose[cryptography]
python-dotenv
sqlalchemy
uvicorn
python-multipart
